fix: the trapdoor thickness is the server's float32
The Bedrock server stores 0.1825 as a float32 (0x14a8262e4); the box now carries that exact value rather than the decimal.

+++ server/block/model/trapdoor.go
@@ -8,7 +8,7 @@ import (
 // trapdoorThickness is the thickness of a trapdoor's collision box as the
 // Bedrock server has it, 2.92 of 16. Java's 3/16 leaves a player who lands
 // on the trapdoor standing inside the box rather than on it.
-const trapdoorThickness = 0.1825
+const trapdoorThickness = float64(float32(0.1825))
 
 // Trapdoor is a model used for trapdoors. It has no solid faces and a bounding box that changes depending on
 // the direction of the trapdoor.
